analytics/requirements.txt
# Core
# python==3.9.21
pandas>=1.3.0,<2.0.0
numpy>=1.20.0,<2.0.0
scikit-learn>=1.0.0,<1.5.0
pyyaml>=6.0,<7.0
joblib>=1.1.0,<1.5.0

# for future extension
### 
# torch>=1.10.0,<2.0.0
# mlflow>=1.20.0,<2.5.0
# psycopg2-binary
